Rename get-account-recovery-password-reset-options to get-account-recovery-reset-options, since I found that even myself couldn't mentally track the former.
